+++ .eslintrc.js
@@ -3,12 +3,58 @@ module.exports = {
     'eslint:recommended',
     'plugin:@typescript-eslint/recommended',
     'plugin:prettier/recommended',
+    'plugin:import/recommended',
+    'plugin:import/typescript',
   ],
   plugins: ['@typescript-eslint'],
-  rules: {
-    '@typescript-eslint/no-explicit-any': 'off',
-  },
+  ignorePatterns: ['dist', '/coverage', 'node_modules'],
   env: {
     node: true,
   },
+  overrides: [
+    {
+      files: ['*.ts'],
+      extends: [
+        'eslint:recommended',
+        'plugin:@typescript-eslint/recommended',
+        'plugin:prettier/recommended',
+        'plugin:import/recommended',
+        'plugin:import/typescript',
+      ],
+      parserOptions: {
+        project: ['tsconfig.eslint.json'],
+      },
+      rules: {
+        '@typescript-eslint/interface-name-prefix': 'off',
+        '@typescript-eslint/explicit-function-return-type': 'off',
+        '@typescript-eslint/explicit-module-boundary-types': 'off',
+        '@typescript-eslint/no-explicit-any': 'off',
+        '@typescript-eslint/ban-ts-comment': 'off',
+        'import/no-cycle': 'error',
+        'import/no-unresolved': 'error',
+        '@typescript-eslint/no-for-in-array': 'error',
+        'no-implied-eval': 'off',
+        '@typescript-eslint/no-implied-eval': 'error',
+        '@typescript-eslint/no-misused-promises': 'error',
+        'require-await': 'off',
+        '@typescript-eslint/require-await': 'error',
+        '@typescript-eslint/restrict-plus-operands': 'error',
+        '@typescript-eslint/unbound-method': 'error',
+      },
+      settings: {
+        'import/parsers': {
+          '@typescript-eslint/parser': ['.ts'],
+        },
+        'import/resolver': {
+          typescript: {},
+        },
+      },
+    },
+    {
+      files: ['examples/**/*.ts'],
+      rules: {
+        'import/no-unresolved': 'off',
+      },
+    },
+  ],
 };