+++ src/app/types.ts
@@ -1,22 +1,29 @@
-import { Module } from '../modules';
-import { AppLogger } from './services';
+import {
+  ModuleInstace as BaseModuleInstance,
+  ModuleMeta as BaseModuleMeta,
+} from '../modules/types';
+import { AppLogger } from './services/logger';
 
 export type AppEvents =
   | 'beforeStart'
   | 'afterStart'
   | 'beforeInit'
   | 'afterInit'
-  | 'beforeHooks'
-  | 'afterHooks'
   | 'beforeStop'
   | 'afterStop';
+
 export type AppEventHandler = () => unknown;
-export type AppModule = Module & Partial<Record<AppEvents, AppEventHandler>>;
-export interface AppModuleConstructor {
+
+export type ModuleInstance = BaseModuleInstance &
+  Partial<Record<AppEvents, AppEventHandler>>;
+
+export interface Module {
   beforeStart?: AppEventHandler;
   beforeInit?: AppEventHandler;
-  new (...args: any): AppModule;
+  new (...args: any): ModuleInstance;
 }
+
+export type ModuleMeta = BaseModuleMeta<Module>;
 
 export type AppOptions = {
   logger?: AppLogger | false;