+++ src/modules/index.ts
@@ -1,3 +1,0 @@
-export * from './types';
-export * from './container';
-export * from './builder';